// types/EcommerceType.ts
export interface ProductListType {
  prikey:number,
  registrationdate: string;
  sphere: string;
  financialyear: string;
  typeofinfo: string;
  fundtype: string;
  vote_code: string;
  vote_name: string;
  sub_subprogramme_code: string;
  sub_subprogramme_name: string;
  service_area_code: string;
  service_area_name: string;
  programme_code: string;
  programme_name: string;
  subprogramme_code: string;
  subprogramme_name: string;
  budget_output_code: string;
  budget_output_description: string;
  project_code: string;
  item_code: string;
  item_description: string;
  fundingsourcecode: string;
  fundingsource: string;
  amount: number

}